Combat
Added basic combat in which players and NPCs have health, defense, and attack power.  Running into an NPC damages its and vice versa. Dead NPCs get rendered first so that living actors and items get priority and appear above them. Also added diagonal movement for the player and a 3rd gamestate for when the player is dead.