Weighted RNG, Stat Changes
Made spawn rates weighted depending on dungeon floor level as opposed to randomly picking numbers between 1 and 100. This means stronger enemies and damaging scrolls are more likely to appear as you  progress further. Also made changes to player's starting stats, along with the stats of enemies and the healing and damaging ratios of useable items.